amsterdam: workaround docker-compose issue
If we fetch latest version of docker-compose via setup.py
installation or pip then we risk an incompatibility issue between
docker and docker-compose because the latest claims it wants the
latest docker engine version.

But docker-compose is lying to us as it can downgrade its protocol
to talk correctly to a lower version of docker engine. To do so
we can pass the COMPOSE_API_VERSION environment variable when
running the docker-compose variable.

This patch is doing that as well as factorizing the definition of
the environment.